requirements.txt
django
pypdf
django-colorfield
django-address
